+++ e2e/bash/test_bash_aot.py
@@ -150,15 +150,24 @@ cmd --option=<FOO>;
 
 
 # https://github.com/adaszko/complgen/issues/41
-@pytest.mark.xfail(reason="Not implemented yet")
 def test_respects_ignore_case_option(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --case-lower | --CASE-UPPER;'''
     with completion_script_path(complgen_binary_path, GRAMMAR) as path:
         input = r'''COMP_WORDS=(cmd --case-); COMP_CWORD=1; bind "set completion-ignore-case on"; _cmd; printf '%s\n' "${COMPREPLY[@]}"'''
-        assert get_sorted_bash_completions(path, input) == sorted(['--case-lower ', '--case-UPPER '])
+        assert get_sorted_bash_completions(path, input) == sorted(['--case-lower', '--CASE-UPPER'])
 
         input = r'''COMP_WORDS=(cmd --case-); COMP_CWORD=1; bind "set completion-ignore-case off"; _cmd; printf '%s\n' "${COMPREPLY[@]}"'''
-        assert get_sorted_bash_completions(path, input) == sorted(['--case-lower '])
+        assert get_sorted_bash_completions(path, input) == sorted(['--case-lower'])
+
+
+def test_respects_ignore_case_option_subwords(complgen_binary_path: Path):
+    GRAMMAR = r'''cmd --option=(lower | UPPER);'''
+    with completion_script_path(complgen_binary_path, GRAMMAR) as path:
+        input = r'''COMP_WORDS=(cmd --option=LO); COMP_CWORD=1; bind "set completion-ignore-case on"; _cmd; printf '%s\n' "${COMPREPLY[@]}"'''
+        assert get_sorted_bash_completions(path, input) == sorted(['lower'])
+
+        input = r'''COMP_WORDS=(cmd --option=LO); COMP_CWORD=1; bind "set completion-ignore-case off"; _cmd; printf '%s\n' "${COMPREPLY[@]}"'''
+        assert get_sorted_bash_completions(path, input) == sorted([''])
 
 
 def test_fallback_completes_default(complgen_binary_path: Path):
@@ -172,7 +181,7 @@ def test_fallbacks_on_no_matches(complgen_binary_path: Path):
     GRAMMAR = r'''cmd (foo || --bar);'''
     with completion_script_path(complgen_binary_path, GRAMMAR) as path:
         input = r'''COMP_WORDS=(cmd --); COMP_CWORD=1; _cmd; printf '%s\n' "${COMPREPLY[@]}"'''
-        assert get_sorted_bash_completions(path, input) == sorted(['--bar '])
+        assert get_sorted_bash_completions(path, input) == sorted(['--bar'])
 
 
 def test_subword_fallback_completes_default(complgen_binary_path: Path):