+++ e2e/bash/test_bash_jit.py
@@ -1,17 +1,13 @@
 import os
-import sys
 import string
 import tempfile
-import subprocess
-import contextlib
 from pathlib import Path
-from typing import Optional, Generator
 
 import pytest
 from hypothesis import given, settings
 from hypothesis.strategies import text
 
-from conftest import complgen_binary_path, set_working_dir, temp_file_with_contents, gen_bash_jit_completions_script_path, get_sorted_bash_completions
+from conftest import complgen_binary_path, set_working_dir, gen_bash_jit_completions_script_path, get_sorted_bash_completions
 from common import LSOF_FILTER_GRAMMAR, STRACE_EXPR_GRAMMAR
 
 
@@ -39,7 +35,7 @@ cargo test --test testname;
 def test_jit_completes_literal_prefix(complgen_binary_path: Path):
     GRAMMAR = '''cmd (--help | --version);'''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--h') as completion_script:
-        input = r'''COMP_WORDS=(cmd --h); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['--help'])
 
 
@@ -91,33 +87,33 @@ cargo +<toolchain>;
 <toolchain> ::= stable-aarch64-apple-darwin | stable-x86_64-apple-darwin;
 '''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='+') as completion_script:
-        input = r'''COMP_WORDS=(cmd +); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['+stable-aarch64-apple-darwin', '+stable-x86_64-apple-darwin'])
 
 
 def test_jit_completes_strace_expr(complgen_binary_path: Path):
     with gen_bash_jit_completions_script_path(complgen_binary_path, STRACE_EXPR_GRAMMAR, ['-e'], prefix='trace=') as completion_script:
-        input = r'''COMP_WORDS=(strace trace=); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(strace); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['!', '%file', 'file', 'all'])
 
 
 def test_jit_completes_lsf_filter(complgen_binary_path: Path):
     with gen_bash_jit_completions_script_path(complgen_binary_path, LSOF_FILTER_GRAMMAR, prefix='-sTCP:') as completion_script:
-        input = r'''COMP_WORDS=(lsf -sTCP:); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(lsf); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['^', 'LISTEN', 'CLOSED'])
 
 
 def test_jit_subword_descriptions(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option=(arg1 "descr1" | arg2 "descr2");'''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=') as completion_script:
-        input = r'''COMP_WORDS=(cmd --option=); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['arg1', 'arg2'])
 
 
 def test_jit_completes_subword_external_command(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option={{{ echo -e "argument\tdescription" }}};'''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=') as completion_script:
-        input = r'''COMP_WORDS=(cmd --option=); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['argument'])
 
 
@@ -128,7 +124,7 @@ cmd --option=<FOO>;
 <FOO@bash> ::= {{{ echo bash }}};
 '''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=') as completion_script:
-        input = r'''COMP_WORDS=(cmd --option=); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['bash'])
 
 
@@ -138,7 +134,7 @@ mygrep --color "use markers to highlight the matching strings"=<WHEN>;
 <WHEN> ::= always | never | auto;
 '''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--color') as completion_script:
-        input = r'''COMP_WORDS=(mygrep --color); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(mygrep); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['--color='])
 
 
@@ -152,33 +148,42 @@ def test_fallback_completes_default(complgen_binary_path: Path):
 def test_fallbacks_on_no_matches(complgen_binary_path: Path):
     GRAMMAR = r'''cmd (foo || --bar);'''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--') as completion_script:
-        input = r'''COMP_WORDS=(cmd --); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['--bar'])
 
 
 def test_subword_fallback_completes_default(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option=(primary || secondary);'''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=') as completion_script:
-        input = r'''COMP_WORDS=(cmd --option=); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['primary'])
 
 
 def test_subword_fallbacks_on_no_matches(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option=(primary || secondary);'''
     with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=sec') as completion_script:
-        input = r'''COMP_WORDS=(cmd --option=sec); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['secondary'])
 
 
-# https://github.com/adaszko/complgen/issues/41
 def test_respects_ignore_case_option(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --case-lower | --CASE-UPPER;'''
-    with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--case-', ignore_case=True) as completion_script:
-        input = r'''COMP_WORDS=(cmd --case-); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+    with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--case-') as completion_script:
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; bind "set completion-ignore-case on"; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['--case-lower', '--CASE-UPPER'])
-    with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--case-', ignore_case=False) as completion_script:
-        input = r'''COMP_WORDS=(cmd --case-); COMP_CWORD=1; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+    with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--case-') as completion_script:
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; bind "set completion-ignore-case off"; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
         assert get_sorted_bash_completions(completion_script, input) == sorted(['--case-lower'])
+
+
+def test_respects_ignore_case_option_subwords(complgen_binary_path: Path):
+    GRAMMAR = r'''cmd --option=(lower | UPPER);'''
+    with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=LO') as completion_script:
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; bind "set completion-ignore-case on"; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        assert get_sorted_bash_completions(completion_script, input) == sorted(['lower'])
+    with gen_bash_jit_completions_script_path(complgen_binary_path, GRAMMAR, prefix='--option=LO') as completion_script:
+        input = r'''COMP_WORDS=(cmd); COMP_CWORD=1; bind "set completion-ignore-case off"; __complgen_jit; printf '%s\n' "${COMPREPLY[@]}"'''
+        assert get_sorted_bash_completions(completion_script, input) == sorted([''])
 
 
 def test_funky_spec_command_name(complgen_binary_path: Path):