+++ src/jit/bash.rs
@@ -3,15 +3,23 @@ use std::io::Write;
 use slice_group_by::GroupBy;
 use ustr::Ustr;
 
-use crate::{aot::bash::{make_id_from_command_map, make_string_constant, write_generic_subword_fn, write_subword_fn}, dfa::DFA, grammar::Specialization, regex::Input};
+use crate::{
+    aot::bash::{
+        make_id_from_command_map, make_string_constant, write_generic_subword_fn, write_match_fn,
+        write_subword_fn, MATCH_FN_NAME,
+    },
+    dfa::DFA,
+    grammar::Specialization,
+    regex::Input,
+};
 
-use super::{get_subword_transitions, get_transitions, make_external_command_fn_name, make_subword_fn_name};
-
+use super::{
+    get_subword_transitions, get_transitions, make_external_command_fn_name, make_subword_fn_name,
+};
 
 pub fn write_bash_completion_shell_code<W: Write>(
     completed_command: &str,
     dfa: &DFA,
-    ignore_case: bool,
     words_before_cursor: &[&str],
     entered_prefix: &str,
     superfluous_prefix: &str,
@@ -23,13 +31,19 @@ pub fn write_bash_completion_shell_code<W: Write>(
     let mut transitions = get_transitions(&dfa, &words_before_cursor);
     transitions.sort_unstable_by_key(|input| input.get_fallback_level());
 
+    write_match_fn(output)?;
+
     let id_from_cmd = make_id_from_command_map(dfa);
 
     for (cmd, id) in &id_from_cmd {
-        write!(output, r#"{} () {{
+        write!(
+            output,
+            r#"{} () {{
     {cmd}
 }}
-"#, make_external_command_fn_name(completed_command, *id))?;
+"#,
+            make_external_command_fn_name(completed_command, *id)
+        )?;
     }
 
     let id_from_dfa = get_subword_transitions(&transitions);
@@ -43,30 +57,59 @@ pub fn write_bash_completion_shell_code<W: Write>(
 
     // Generate shell code for fallback levels in order, optionally calling shell functions defined
     // above.
-    writeln!(output, r#"__complgen_jit () {{
+    writeln!(
+        output,
+        r#"__complgen_jit () {{
     local -a completions=()
-    local -a matches=()
-"#)?;
+    local -a matches=()"#
+    )?;
 
-    let groups: Vec<&[Input]> = transitions.linear_group_by_key(|t| t.get_fallback_level()).collect();
+    if !entered_prefix.is_empty() {
+        writeln!(
+            output,
+            r#"    local ignore_case=$(bind -v | grep completion-ignore-case | cut -d' ' -f3)"#
+        )?;
+    }
+
+    let groups: Vec<&[Input]> = transitions
+        .linear_group_by_key(|t| t.get_fallback_level())
+        .collect();
     for group in groups {
-        let literals: Vec<&Ustr> = group.iter().filter_map(|t| match t {
-            Input::Literal(literal, ..) => Some(literal),
-            _ => None,
-        }).collect();
+        let literals: Vec<&Ustr> = group
+            .iter()
+            .filter_map(|t| match t {
+                Input::Literal(literal, ..) => Some(literal),
+                _ => None,
+            })
+            .collect();
         if !literals.is_empty() {
-            writeln!(output, r#"    completions+=({})"#, itertools::join(literals.iter().map(|s| make_string_constant(s)), " "))?;
+            writeln!(
+                output,
+                r#"    completions+=({})"#,
+                itertools::join(literals.iter().map(|s| make_string_constant(s)), " ")
+            )?;
         }
 
         // An external command -- execute it and collect stdout lines as completions
         for cmd in group.iter().filter_map(|t| match t {
             Input::Command(cmd, ..) => Some(*cmd),
-            Input::Nonterminal(_, Some(Specialization { bash: None, generic: Some(cmd), .. }), ..) => Some(*cmd),
+            Input::Nonterminal(
+                _,
+                Some(Specialization {
+                    bash: None,
+                    generic: Some(cmd),
+                    ..
+                }),
+                ..,
+            ) => Some(*cmd),
             _ => None,
         }) {
             let command_id = id_from_cmd.get(&cmd).unwrap();
             let fn_name = make_external_command_fn_name(completed_command, *command_id);
-            writeln!(output, r#"    readarray -t -O ${{#completions[@]}} completions < <({fn_name} {prefix_constant})"#)?
+            writeln!(
+                output,
+                r#"    readarray -t -O ${{#completions[@]}} completions < <({fn_name} {prefix_constant})"#
+            )?
         }
 
         for subdfa in group.iter().filter_map(|t| match t {
@@ -75,48 +118,57 @@ pub fn write_bash_completion_shell_code<W: Write>(
         }) {
             let subdfa_id = id_from_dfa.get(subdfa).unwrap();
             let fn_name = make_subword_fn_name(completed_command, *subdfa_id);
-            writeln!(output, r#"    readarray -t -O ${{#completions[@]}} completions < <({fn_name} complete {prefix_constant})"#)?
+            writeln!(
+                output,
+                r#"    readarray -t -O ${{#completions[@]}} completions < <({fn_name} complete {prefix_constant})"#
+            )?
         }
         for cmd in group.iter().filter_map(|t| match t {
-            Input::Nonterminal(_, Some(Specialization { bash: Some(cmd), .. }), ..) => Some(*cmd),
+            Input::Nonterminal(
+                _,
+                Some(Specialization {
+                    bash: Some(cmd), ..
+                }),
+                ..,
+            ) => Some(*cmd),
             _ => None,
         }) {
             let command_id = id_from_cmd.get(&cmd).unwrap();
             let fn_name = make_external_command_fn_name(completed_command, *command_id);
-            writeln!(output, r#"    readarray -t -O ${{#completions[@]}} completions < <({fn_name} {prefix_constant})"#)?
+            writeln!(
+                output,
+                r#"    readarray -t -O ${{#completions[@]}} completions < <({fn_name} {prefix_constant})"#
+            )?
         }
 
         if entered_prefix.is_empty() {
             writeln!(output, r#"    matches=("${{completions[@]}}")"#)?;
-        }
-        else {
+        } else {
             // Filter `completions` by `entered_prefix`, respecting `ignore_case`
-            if ignore_case {
-                writeln!(output, r#"    local prefix={prefix_constant}
-        lowercase_prefix=${{prefix,,}}
-        for item in "${{completions[@]}}"; do
-            [[ "${{item,,}}" = ${{lowercase_prefix}}* ]] && matches+=("$item")
-        done
-"#)?;
-            } else {
-                writeln!(output, r#"    for item in "${{completions[@]}}"; do
-        [[ $item = {prefix_constant}* ]] && matches+=("$item")
-    done"#)?;
-            }
+            writeln!(
+                output,
+                r#"    readarray -t matches < <(printf "%s\n" "${{completions[@]}}" | {MATCH_FN_NAME} "$ignore_case" {prefix_constant})"#
+            )?;
         }
 
         // Strip `superfluous_prefix` bashism
         if superfluous_prefix.is_empty() {
-            writeln!(output, r#"    if [[ ${{#matches[@]}} -gt 0 ]]; then
+            writeln!(
+                output,
+                r#"    if [[ ${{#matches[@]}} -gt 0 ]]; then
         COMPREPLY=("${{matches[@]}}")
         return
-    fi"#)?;
+    fi"#
+            )?;
         } else {
-            writeln!(output, r#"    if [[ ${{#matches[@]}} -gt 0 ]]; then
+            writeln!(
+                output,
+                r#"    if [[ ${{#matches[@]}} -gt 0 ]]; then
         matches=("${{matches[@]#{superfluous_prefix}}}")
         COMPREPLY=("${{matches[@]}}")
         return
-    fi"#)?;
+    fi"#
+            )?;
         }
     }
     writeln!(output, r#"}}"#)?;